backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.1.3
asyncpg==0.27.0
python-dotenv==1.0.0

# dev/test
pytest==7.4.0
httpx==0.24.1
pandas==2.2.2
openpyxl==3.1.2
